heredoc関連の関数をcmd_heredoc_funcs.cに入れた
