+++ Makefile
@@ -22,7 +22,8 @@ HEADER_FILES = builtin.h env.h lexer.h parse.h	\
 	execution.h minishell.h path.h utils.h
 
 SRCS = cmd_cmd_invocation.c cmd_cmd_invocation2.c cmd_exec_command.c		\
-	cmd_exec_commands.c cmd_pipe.c cmd_redirection.c convert_ast2cmdinvo.c	\
+	cmd_exec_commands.c cmd_pipe.c cmd_redirection.c cmd_heredoc_funcs.c	\
+	convert_ast2cmdinvo.c													\
 	cmd_status.c signal.c env_setter.c minishell_error_msg.c env.c exec.c	\
 	cmd_exec_builtin.c														\
 																			\
